Finishing adding remove_remote tests

+++ tests/test_add_remote_catalog.py
@@ -19,8 +19,7 @@ def dummy_remote_add(catalog_url: str, catalog_name: str, path: Path):
     library_already_exists = exists(join(path, 'library.yaml'))
     if not library_already_exists:
         with open(join(path, catalog_name, 'library.yaml'), 'w') as f:
-            catalogs = {'catalogs': []}
-            catalogs['catalogs'].append(catalog_url)
+            catalogs = {'catalogs': [catalog_url]}
             f.seek(0)
             yaml.dump(catalogs, f)
             symlink(join(path, catalog_name, 'library.yaml'), join(path, 'library.yaml'))
@@ -32,11 +31,19 @@ def dummy_remote_add(catalog_url: str, catalog_name: str, path: Path):
             f.seek(0)
             yaml.dump(catalogs, f)
 
-    with open(join(path, 'index.yaml'), '+a') as f:
-        catalogs = yaml.safe_load(f) if yaml.safe_load(f) is not None else {'catalogs': []}
-        catalogs['catalogs'].append(catalog_name)
-        f.seek(0)
-        yaml.dump(catalogs, f)
+    index_already_exists = exists(join(path, 'index.yaml'))
+    if not index_already_exists:
+        with open(join(path, 'index.yaml'), 'w') as f:
+            catalogs = {'catalogs': [catalog_name]}
+            f.seek(0)
+            yaml.dump(catalogs, f)
+    else:
+        with open(join(path, 'index.yaml'), 'r+') as f:
+            existing_catalogs = yaml.safe_load(f)
+            catalogs = existing_catalogs if existing_catalogs is not None else {'catalogs': []}
+            catalogs['catalogs'].append(catalog_name)
+            f.seek(0)
+            yaml.dump(catalogs, f)
                 
     # create the json file
     with open(join(path, catalog_name + '.json'), 'a') as f:
